Add Zones support
This commit add the necessary pieces to support controlling Zone
accutators that are often found in ducted heat pump units.

The magic for how the zones are controlled was documented in an
issue comment here:
  https://github.com/SwiCago/HeatPump/issues/39
which mentions the packet type of 0x15, which seems to be not
documented anywhere else.

These units support a configuration of either 4 or 8 zones.